+++ backend/footprint/src/main/java/foot/footprint/domain/article/application/create/CreateArticleService.java
@@ -1,0 +1,8 @@
+package foot.footprint.domain.article.application.create;
+
+import foot.footprint.domain.article.dto.CreateArticleRequest;
+
+public interface CreateArticleService {
+
+    public Long create(CreateArticleRequest request, Long memberId);
+}
